fix: split 'json to typebox' process in different files
