+++ src/Json/getRangeEnumeration.ts
@@ -1,0 +1,31 @@
+// TODO: add description
+/**
+ *
+ * @param rangeEnumeration
+ * @returns
+ */
+export const getRangeEnumeration = (
+  rangeEnumeration: string
+): [...(number | null)[]] | null => {
+  //TODO: define if null or undefine
+  if (rangeEnumeration.length === 0) return null; // empty string case
+
+  if (rangeEnumeration.includes("..")) {
+    const minAndMax = rangeEnumeration.split("..");
+
+    const minimum = minAndMax[0]
+      ? Number(minAndMax[0].replace(/\D/g, ""))
+      : null;
+    const maximum = minAndMax[1]
+      ? Number(minAndMax[1].replace(/\D/g, ""))
+      : null;
+    return [minimum, maximum];
+  }
+
+  if (rangeEnumeration.split(",").length > 1)
+    return rangeEnumeration
+      .split(",")
+      .map((element: string) => Number(element));
+
+  return null; // no valid case
+};
